requirements.txt
alembic==1.3.3
biopython==1.76
certifi==2019.11.28
chardet==3.0.4
click==6.7
colorama==0.4.3
docopt==0.6.2
feedparser==5.2.1
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gitsome==0.8.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.0
Mako==1.1.1
MarkupSafe==1.1.1
numpy==1.18.1
ply==3.11
prompt-toolkit==2.0.10
Pygments==2.5.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
pytz==2016.10
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
uritemplate==3.0.1
uritemplate.py==3.0.2
urllib3==1.25.8
wcwidth==0.1.8
Werkzeug==0.16.1
WTForms==2.2.1
